@@ -4,19 +4,19 @@
   "homepage": "./",
   "license": "MIT",
   "dependencies": {
-    "@apollo/client": "^3.3.15",
+    "@apollo/client": "^3.5.10",
     "@apollo/react-hooks": "^4.0.0",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.1",
-    "@testing-library/user-event": "^12.2.0",
-    "graphql": "^15.5.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
-    "redux": "^4.0.5",
-    "web-vitals": "^0.2.4"
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.4",
+    "graphql": "^16.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
+    "redux": "^4.1.2",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -43,7 +43,7 @@
     ]
   },
   "devDependencies": {
-    "node-sass": "^4.0.0",
+    "sass": "^1.49.11",
     "redux-devtools-extension": "^2.13.9"
   }
 }
